feat(page-header): migrates to material ui
Migrates the PageHeader component to material ui

re #143
