+++ pages/players/index.tsx
@@ -1,5 +1,5 @@
 import Head from 'next/head';
-import PageHeader from '@components/layout/PageHeader';
+import PageHeader from '@components/organisms/PageHeader';
 import { Grid, GridItem } from '@chakra-ui/react';
 import { PlayerSearch } from '@components/players/PlayerSearch';
